+++ crates/ryft-core/CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - Release Date
 
+### Changed
+
+- Significantly enhanced the `Parameterized` trait.
+
 ## [0.0.1] - 2026-02-22
 
 ### Added