+++ crates/ryft-macros/CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - Release Date
 
+### Changed
+
+- Updated the `#[derive(Parameterized)]` macro to be compatible with the updated trait in `ryft-core`.
+
 ## [0.0.1] - 2026-02-22
 
 ### Added